refactor: use mixin for dynamic properties
