+++ TM1py/Objects/MDXView.py
@@ -5,28 +5,29 @@ import json
 import re
 from typing import Dict, Optional
 
+from TM1py.Objects.DynamicPropertiesMixin import DynamicPropertiesMixin
 from TM1py.Objects.View import View
 from TM1py.Utils import case_and_space_insensitive_equals
 
-MDX_VIEW_EXCLUDED_KEYS = frozenset(
-    {
-        "@odata.type",
-        "@odata.context",
-        "@odata.etag",
-        "Name",
-        "MDX",
-        "Cube",
-        "Attributes",
-        "LocalizedAttributes",
-    }
-)
 
-
-class MDXView(View):
+class MDXView(DynamicPropertiesMixin, View):
     """Abstraction on TM1 MDX view
 
     IMPORTANT. MDXViews can't be seen through the old TM1 clients (Architect, Perspectives). They do exist though!
     """
+
+    _DYNAMIC_PROPERTIES_EXCLUDED_KEYS = frozenset(
+        {
+            "@odata.type",
+            "@odata.context",
+            "@odata.etag",
+            "Name",
+            "MDX",
+            "Cube",
+            "Attributes",
+            "LocalizedAttributes",
+        }
+    )
 
     def __init__(self, cube_name: str, view_name: str, MDX: str, dynamic_properties: Optional[Dict] = None):
         View.__init__(self, cube_name, view_name)
@@ -49,14 +50,6 @@ class MDXView(View):
     @MDX.setter
     def MDX(self, value: str):
         self._mdx = value
-
-    @property
-    def dynamic_properties(self) -> Dict:
-        return self._dynamic_properties
-
-    @dynamic_properties.setter
-    def dynamic_properties(self, value: Optional[Dict]) -> None:
-        self._dynamic_properties = value or {}
 
     @property
     def body(self) -> str:
@@ -97,7 +90,7 @@ class MDXView(View):
             cube_name=view_as_dict["Cube"]["Name"] if not cube_name else cube_name,
             view_name=view_as_dict["Name"],
             MDX=view_as_dict["MDX"],
-            dynamic_properties={k: v for k, v in view_as_dict.items() if k not in MDX_VIEW_EXCLUDED_KEYS},
+            dynamic_properties=cls._filter_dynamic_properties(view_as_dict),
         )
 
     def construct_body(self) -> str:
@@ -105,5 +98,5 @@ class MDXView(View):
         mdx_view_as_dict["@odata.type"] = "ibm.tm1.api.v1.MDXView"
         mdx_view_as_dict["Name"] = self._name
         mdx_view_as_dict["MDX"] = self._mdx
-        mdx_view_as_dict.update({k: v for k, v in self._dynamic_properties.items() if k not in MDX_VIEW_EXCLUDED_KEYS})
+        mdx_view_as_dict.update(self._filter_dynamic_properties(self._dynamic_properties))
         return json.dumps(mdx_view_as_dict, ensure_ascii=False)
